requirements.txt
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cytoolz==0.11.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
lru-dict==1.1.6
MarkupSafe==1.1.1
multiaddr==0.0.9
nbformat==5.1.2
netaddr==0.8.0
numpy==1.19.5
pandas==1.2.1
parsimonious==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.14.3
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pycryptodome==3.9.9
Pygments==2.7.4
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pyzmq==22.0.2
requests==2.25.1
retrying==1.3.3
rlp==2.0.1
six==1.15.0
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
varint==1.0.2
wcwidth==0.2.5
web3==5.15.0
websockets==8.1
Werkzeug==1.0.1
